Create The Minion Game
HackerRank's Minion Game where the user enters a string, and substrings are counted based on whether the letter is a vowel or a consonant. Stuart gets a point for every substring that starts with a consonant, and Kevin gets a point for every substring that starts with a vowel. Let's see who wins, shall we?